0034.find-first-and-last-position-of-element-in-sorted-array.ts
function searchRange(nums: number[], target: number): number[] {
  let left = 0;
  let right = nums.length - 1;
  let f = -1;
  let l = -1;
  while (left <= right) {
    const mid = Math.floor((left + right) / 2);
    if (nums[mid] === target) {
      f = mid;
    }
    if (nums[mid] < target) {
      left = mid + 1;
    } else {
      right = mid - 1;
    }
  }
  right = nums.length - 1;
  while (left <= right) {
    const mid = Math.floor((left + right) / 2);
    if (nums[mid] === target) {
      l = mid;
    }
    if (nums[mid] <= target) {
      left = mid + 1;
    } else {
      right = mid - 1;
    }
  }
  return [f, l];
}

export default searchRange;
